lib: lite_buttons: add range to num buttons and reduce maximum
Add range to CONFIG_LITE_BUTTONS_NUM_PINS and reduce
the maximum size to 32. Unrealistic to have 64 pins used for buttons.
As a consequence, reduce size/type of internal variable
global.pin_active to uint32_t from uint64_t.

+++ lib/lite_buttons/lite_buttons.c
@@ -138,7 +138,7 @@ struct lite_buttons_state {
 	uint32_t detection_delay;
 	struct lite_timer timer;
 	uint8_t pin_states[((NUM_PINS + 1) * BITS_PER_PIN) / 8];
-	uint64_t pin_active;
+	uint32_t pin_active;
 };
 
 static struct lite_buttons_state global;
